requirements.txt
quart               # Основной фреймворк
bcrypt              # Для хэширования
python-dotenv       # Для переменных окружения
psycopg2-binary     # Для PostgreSQL
asyncpg             # Для асинхронной PostgreSQL
alembic             # Для миграций
psycopg2-binary
Faker
tqdm
redis>=4.5.0